processing/requirements.txt
# --- Google APIs for Drive access ---
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

# --- OpenAI API ---
openai==1.102.0

# --- SendGrid for emailing ---
sendgrid==6.12.4
python-http-client==3.3.7

# --- Data validation ---
pydantic==2.11.7

# --- Utilities ---
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
